backend/pyproject.toml
[tool.poetry]
name = "all-txt backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "all" }]

[tool.poetry.scripts]
alembic-migrate = "alembic revision --revision"
alembic-upgrade = "alembic upgrade head"
alembic-downgrade = "alembic downgrade -1"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
pydantic-settings = "^2.2.1"
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.30"
psycopg = {extras = ["binary"], version = "^3.1.19"}
alembic = "^1.13.1"
pyyaml = "^6.0.1"
fastapi-limiter = "^0.1.6"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}


[tool.poetry.group.dev.dependencies]
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
